requirement.txt

GitPython
numpy
numexpr
google-cloud-storage
parallel-sort
torch 
torchvision
pytorch-transformers
huggingface-hub
safetensors
tokenizers
importlib-metadata
transformers
scikit-learn
pandas
dask
pyarrow